feat(katana): default to deserialize response as JSON (#157)
Deserialize the JSON-RPC response object as a raw JSON to make the client compatible with other Starknet RPC spec. 

Currently, we deserialize the response as concrete types from `starknet-rs` crate. The crate only supports a single RPC spec at a time so the types are bounded to a single RPC spec. Meaning the client is only compatible when interacting the a server of a similar spec. 

Deserializing as raw JSON makes the client more flexible and can be used to interact with Starknet servers regardless of the RPC spec version.